backend/requirements.txt
alembic==1.8.1
astroid==2.11.6
atomicwrites==1.4.0
attrs==21.4.0
autopep8==1.6.0
bcrypt==4.0.0
cffi==1.15.1
click==8.1.3
colorama==0.4.5
coverage==6.4.1
cryptography==38.0.1
dill==0.3.5.1
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
importlib-metadata==4.12.0
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.7.1
Mako==1.2.2
MarkupSafe==2.1.1
marshmallow==3.17.1
marshmallow-sqlalchemy==0.28.1
mccabe==0.7.0
mypy==0.971
mypy-extensions==0.4.3
mysqlclient==2.1.1
packaging==21.3
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.4.0
pylint==2.14.3
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-plugin-utils==0.7
PyMuPDF==1.20.2
pyparsing==3.0.9
pytest==7.1.2
pytest-cov==3.0.0
python-dotenv==0.20.0
six==1.16.0
SQLAlchemy==1.4.39
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.0
types-Flask-SQLAlchemy==2.5.9
types-SQLAlchemy==1.4.51
typing_extensions==4.2.0
Werkzeug==2.1.2
wrapt==1.14.1
WTForms==3.0.1
zipp==3.8.0
